Create Q001
This is the solution to problem number 1. 

If we list all the natural numbers below 10 that are multiples of 3 or 5, we get 3, 5, 6 and 9. The sum of these multiples is 23.

Find the sum of all the multiples of 3 or 5 below 1000.

So here we loop through the values from 3 to 999(both inclusive) and add all the values that are either divisible by 3 or 5  and print the value.