Names scores
Using names array containing over five-thousand first names, begin by sorting it into alphabetical order. Then working out the alphabetical value for each name, multiply this value by its alphabetical position in the list to obtain a name score.

For example, when the list is sorted into alphabetical order, COLIN, which is worth 3 + 15 + 12 + 9 + 14 = 53, is the 938th name in the list. So, COLIN would obtain a score of 938 × 53 = 49714.

What is the total of all the name scores in the file?

Ans : 871198282

Here we loop through an array of strings. each element in the array has a name. we need to iterate through each char in the each element and find it's numerical value. That's what we are doing using the charAt function of string in java.